requirements.txt
asgiref==3.9.1
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==2.0.12
Django==5.2.7
django-active-link==0.1.8
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-colorfield==0.14.0
django-js-asset==3.1.2
django-render-partial==0.4
django-widget-tweaks==1.5.0
idna==3.10
pillow==11.3.0
PyMySQL==1.0.2
python-decouple==3.5
pytz==2025.2
requests==2.26.0
setuptools==80.9.0
sqlparse==0.4.2
tzdata==2025.2
urllib3==1.26.20
webencodings==0.5.1
